Move print legend creation in its own file
